made a new channel uri. removed unnecessary <stack> from app/page.tsx
